+++ SomnusLoadBalancer/src/main/java/com/somnus/server/somnuslb/mirrorStatus/domain/MirrorState.java
@@ -1,4 +1,4 @@
-package com.somnus.server.somnuslb.mirrorHealth.domain;
+package com.somnus.server.somnuslb.mirrorStatus.domain;
 
 public enum MirrorState {
     NEW_MIRROR,