+++ SomnusLoadBalancer/src/main/java/com/somnus/server/somnuslb/mirrors/MirrorRegisterController.java
@@ -1,5 +1,7 @@
 package com.somnus.server.somnuslb.mirrors;
 
+import com.somnus.server.somnuslb.mirrors.dto.NewMirrorRegisterDTO;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -8,8 +10,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 @RequestMapping("/mirrors")
 public class MirrorRegisterController {
 
-    @GetMapping("/register")
-    public void registerMirror() {
+    @Autowired
+    private MirrorService mirrorService;
 
+    @GetMapping("/register")
+    public void registerMirror(NewMirrorRegisterDTO mirrorRegisterDTO) {
+        // TODO: Security checks (e.g with token)
+
+        this.mirrorService.addMirror(mirrorRegisterDTO);
     }
 }